requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
idna==3.10
psutil==7.0.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
TgCrypto==1.2.5
typing_extensions==4.12.2
urllib3==2.3.0
